// src/app/models/sjcs/BusquedaGeneralSJCSItem.ts
export class BusquedaGeneralSJCSItem {
  idInstitucion: string;
  nombre: string;
  idTurno: string;
  fechaEstadoColegial: Date;
  idGuardia: string;
  idEstado: string;
  descripcion: string;
  historico: boolean;
  nif: string;
  apellidos: string;
  colegios: any[];
  nColegiado: string;
  inscritoturno: string;
  inscritoguardia: string;
  guardiasPendientes: string;
  telefono: string;
  abreviatura: string;
  constructor() { }
}